requirements.txt
-r requirements/common.txt
-r requirements/base.txt
azure-storage-blob>=12.19.0
django-storages[azure]>=1.14.0
azure-identity>=1.15.0